backend/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "jupyterlab>=4.5.6",
    "owid>=0.0.1",
    "pydantic-ai>=1.70.0",
    "pydantic-ai-slim[openrouter]>=1.70.0",
    "python-dotenv>=1.2.2",
]
